Move and rename classes.  Update README.
