+++ README.md
@@ -1,5 +1,19 @@
-XenoBrainz
-==========
+# A bare bones, BSD Licensed MusicBrainz WS API in Java that is Android compatible.
+
+# Usage
+
+See unit tests.
+
+# Based on work of ssuukk at https://github.com/ssuukk/XenoBrainz
+
+Changes
+* Made code more idiomatic.
+* Added gradle build files.
+* Added a few basic unit tests.
+
+# Original Notice
+
+## XenoBrainz
 
 MusicBrainz Java API, no strings attached
 
@@ -7,3 +21,4 @@ Licensed under BSD License
 Copyright 2012 ssuukk
 
 XenoBrainz is a part of XenoAmp - a slightly different music player for Android
+
