--- src/main/java/pl/qus/xenoamp/musicbrainz/MBDiscList.java
@@ -1,10 +1,0 @@
-package pl.qus.xenoamp.musicbrainz;
-
-import org.jdom2.Element;
-
-public class MBDiscList {
-
-    public MBDiscList(Element e) {
-        // TODO: zakodować
-    }
-}
