+++ src/main/java/pl/qus/xenoamp/musicbrainz/model/MBArtistCredit.java
@@ -1,9 +1,9 @@
-package pl.qus.xenoamp.musicbrainz;
+package pl.qus.xenoamp.musicbrainz.model;
 
 import org.jdom2.Element;
 
 public class MBArtistCredit extends MBNameCredit {
     public MBArtistCredit(Element e) {
-        super(e);
+        super(e.getChild("name-credit", e.getNamespace()));
     }
 }
